Relocate implementations of the public API to single files.
When the `rules_bison` project was first started, the recommended
structure for Bazel rulesets was to have most (or all) of the public
API exposed in a single entry-point file (here, `//bison:bison.bzl`).

Nowadays it's considered better to have the public API in multiple
files, such that the `load()` graph for any given public symbol can be
minimized.

The README will continue to reference the unified `//bison:bison.bzl`
since it's a lot easier to read than a bunch of lengthy `load()` calls.

GitOrigin-RevId: 2eb6df9451ab8faf1e1e7c3b99315a428c86545b